added scss loader
