+++ src/index.js
@@ -1,0 +1,4 @@
+import './main.scss';
+import run from './js/init.js';
+
+run();
