--- src/main.css
@@ -1,110 +1,0 @@
-.page-header {
-  background-image: url("./assets/homepage.jpg");
-  background-repeat: no-repeat;
-  background-size: 100% 100%;
-  background-position: top center;
-  background-attachment: fixed;
-  width: 100%;
-  height: 100vh;
-  margin: 0;
-}
-
-.logo-header {
-  background-image: url("./assets/logohome.png");
-  background-position: center;
-  background-size: contain;
-  width: 40%;
-  background-repeat: no-repeat;
-  height: 200px;
-  margin: auto;
-  position: absolute;
-  top: 1%;
-  right: 32%;
-}
-
-.arrow {
-  text-align: center;
-  margin: 0;
-  position: absolute;
-  bottom: 15%;
-  left: 50%;
-}
-
-.arrow-property {
-  border: solid whitesmoke;
-  border-top-width: medium;
-  border-right-width: medium;
-  border-bottom-width: medium;
-  border-left-width: medium;
-  border-width: 0 3px 3px 0;
-  display: inline-block;
-  padding: 3px;
-  -webkit-transform: rotate(45deg);
-          transform: rotate(45deg);
-}
-
-.bounce {
-  -webkit-animation: bounce 2s infinite;
-          animation: bounce 2s infinite;
-}
-
-@-webkit-keyframes bounce {
-  0%, 20%, 50%, 80%, 100% {
-    -webkit-transform: translateY(0);
-            transform: translateY(0);
-  }
-  40% {
-    -webkit-transform: translateY(-30px);
-            transform: translateY(-30px);
-  }
-  60% {
-    -webkit-transform: translateY(-15px);
-            transform: translateY(-15px);
-  }
-}
-
-@keyframes bounce {
-  0%, 20%, 50%, 80%, 100% {
-    -webkit-transform: translateY(0);
-            transform: translateY(0);
-  }
-  40% {
-    -webkit-transform: translateY(-30px);
-            transform: translateY(-30px);
-  }
-  60% {
-    -webkit-transform: translateY(-15px);
-            transform: translateY(-15px);
-  }
-}
-
-.menu-item {
-  padding: 1em;
-  color: #2d2d2d;
-}
-
-li {
-  cursor: pointer;
-}
-
-.about-img {
-  margin: 10px auto;
-  border-radius: 50%;
-  height: 90%;
-  width: 95%;
-}
-
-.menu-image {
-  height: 14rem;
-}
-
-.about-p {
-  height: 100%;
-  width: 100%;
-  padding-top: 13%;
-}
-
-.activeheader {
-  background-color: white;
-}
-/*# sourceMappingURL=main.css.map */
